jwt + cookie login

+++ dao/user/userDao.go
@@ -4,7 +4,7 @@ import "github.com/manumura/golang-app-device/model/user"
 
 // UserDao : interface defining DB operations on user
 type UserDao interface {
-	GetUser(id int64) (usermodel.User, error)
+	GetUser(id int) (usermodel.User, error)
 	GetUserByUsername(username string) (usermodel.User, error)
 	CheckUsernameUnique(username string) (bool, error)
 	Create(user usermodel.User) (usermodel.User, error)
