+++ service/user/userService.go
@@ -6,7 +6,7 @@ import (
 
 // UserService : interface defining services on user
 type UserService interface {
-	GetUser(id int64) (usermodel.User, error)
+	GetUser(id int) (usermodel.User, error)
 	GetUserByUsername(username string) (usermodel.User, error)
 	CheckUsernameUnique(username string) (bool, error)
 	Create(user usermodel.User) (usermodel.User, error)
